Made dictionary iterations compatible with python 2 and 3
This mainly consists of replacing, for example, dict.iteritems with dict.items.

In the cases where a list of keys is being generated to get the first item from
the list, this usage has been replaced by converting the dictionary to an
iterator and getting the next item from the iterator.
This also means that we aren't generating a whole list just to get a single
item.