Made file reading and writing compatible with python 3
Mostly this is opening files in text mode rather than binary mode, so that
inputs and outputs are not 'bytes' and remain strings.
Where subprocess output is compared with file contents, strings, or type
checked, 'universal_newlines' has been turned on so that it also outputs a
string rather than a 'bytes'.